backend/assisstant/requirements.txt
langchain
google-generativeai
langchain-google-genai
langgraph
langchain-google-vertexai
google-cloud-texttospeech